feat(fe) moving wallet components
